ai-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
spacy==3.7.2
scikit-learn==1.3.2
pdfplumber==0.10.3
python-docx==1.1.0
numpy==1.26.2